+++ src/main/java/hr/dlatecki/algorithms/gen_alg/codecs/DoubleArrayToGrayBinaryCodec.java
@@ -1,0 +1,6 @@
+package hr.dlatecki.algorithms.gen_alg.codecs;
+
+
+public class DoubleArrayToGrayBinaryCodec {
+
+}